PPU refactor (#29)
Initial intent was to implement the full FIFO. But after finding that it's still
pretty undocumented and only enables one single released game and a
few demos to work flawlessly, while giving a substantial performance
drop in the general case, I decided that I better improve the existing
PPU implementation. In the end, I decided to rewrite the PPU code
from scratch.

Neither the old nor the new implementation is anywhere close to pass
the ACID2 test, and the FIFO implementation just barely manage to
show the Nintendo logo at boot. But I recon this could be a good time
to merge the work anyway. The previous PPU implementation lcd.rs,
and the fifo implementation (ppu_fifo) will be dropped soon.

+++ src/apu/noise_gen.rs
@@ -98,7 +98,7 @@ impl NoiseSoundGenerator {
         self.dac = DAC::new();
     }
 
-    pub fn read_reg(&self, address: u16) -> u8 {
+    pub fn read_reg(&self, address: usize) -> u8 {
         match address {
             NR40_REG => 0xFF,
             NR41_REG => 0xFF,
@@ -125,7 +125,7 @@ impl NoiseSoundGenerator {
         }
     }
 
-    pub fn write_reg(&mut self, address: u16, value: u8, seq_step: u8, powered_on: bool) {
+    pub fn write_reg(&mut self, address: usize, value: u8, seq_step: u8, powered_on: bool) {
         // If unpowered, all writes should be ignored except
         // length value if the machine is original Gameboy DMG
         if !powered_on {
