+++ src/apu/square_gen.rs
@@ -143,7 +143,7 @@ impl SquareWaveSoundGenerator {
         }
     }
 
-    pub fn read_reg(&self, address: u16) -> u8 {
+    pub fn read_reg(&self, address: usize) -> u8 {
         match address {
             NR10_REG | NR20_REG => match self.sweep {
                 Some(ref sweep) => sweep.read_reg_nr10() | 0x80,
@@ -173,7 +173,7 @@ impl SquareWaveSoundGenerator {
         }
     }
 
-    pub fn write_reg(&mut self, address: u16, value: u8, seq_step: u8, powered_on: bool) {
+    pub fn write_reg(&mut self, address: usize, value: u8, seq_step: u8, powered_on: bool) {
         // If unpowered, all writes should be ignored except
         // length value if the machine is original Gameboy DMG
         if !powered_on {
